Using drawer instead of menu when clicking hamburger button
